requirements.txt
# ----------------------------
# Telegram bot
# ----------------------------
python-telegram-bot==21.6

# ----------------------------
# OpenAI (OCR/vision -> JSON)
# ----------------------------
openai==1.40.6

# ----------------------------
# App config / utils
# ----------------------------
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ----------------------------
# Logging / consola prolija
# ----------------------------
rich==13.8.1
loguru==0.7.2

# ----------------------------
# PDF generation (facturas servicio)
# HTML/CSS -> PDF
# ----------------------------
weasyprint==62.3
jinja2==3.1.4

# ----------------------------
# Image utilities (descargas, normalización)
# ----------------------------
Pillow==10.4.0

# ----------------------------
# Streamlit dashboard
# ----------------------------
streamlit==1.37.1
pandas==2.2.2

# ----------------------------
# Export (opcional pero útil)
# ----------------------------
openpyxl==3.1.5

# ----------------------------
# Testing
# ----------------------------
pytest==8.3.2
pytest-cov==5.0.0

httpx==0.27.2

# WeasyPrint PDF backend dependency (pin para evitar bug transform)
pydyf==0.10.0


